passage de latitudes et longitudes de float à double et prise en charge du calcul de distance pour getNearestCourses()

+++ TreasureHuntModel/TreasureHuntModel/src/main/java/treasurehunt/model/StepComposite.java
@@ -19,7 +19,7 @@ public class StepComposite extends Step {
 		nextStepsIds= new HashSet<String>();
 	}
 	
-	public StepComposite(String id, float latitude, float longitude) {
+	public StepComposite(String id, double latitude, double longitude) {
 		this();
 		this.id = id;
 		this.latitude = latitude;
